Changed Model Fields

+++ pro/app/models.py
@@ -3,91 +3,92 @@ from datetime import datetime
 # Create your models here.
 
 class addm(models.Model):
-    FirstNumber = models.IntegerField()
-    SecondNumber = models.IntegerField()
+    FirstNumber = models.FloatField()
+    SecondNumber = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.FirstNumber} & {self.SecondNumber} at {self.Time}'
     
 class subm(models.Model):
-    FirstNumber = models.IntegerField()
-    SecondNumber = models.IntegerField()
+    FirstNumber = models.FloatField()
+    SecondNumber = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.FirstNumber} & {self.SecondNumber} at {self.Time}'
     
 class mulm(models.Model):
-    FirstNumber = models.IntegerField()
-    SecondNumber = models.IntegerField()
+    FirstNumber = models.FloatField()
+    SecondNumber = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.FirstNumber} & {self.SecondNumber} at {self.Time}'
     
 class quom(models.Model):
-    FirstNumber = models.IntegerField()
-    SecondNumber = models.IntegerField()
+    FirstNumber = models.FloatField()
+    SecondNumber = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.FirstNumber} & {self.SecondNumber} at {self.Time}'
     
 class remm(models.Model):
-    FirstNumber = models.IntegerField()
-    SecondNumber = models.IntegerField()
+    FirstNumber = models.FloatField()
+    SecondNumber = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.FirstNumber} & {self.SecondNumber} at {self.Time}'
     
 class expm(models.Model):
-    FirstNumber = models.IntegerField()
-    SecondNumber = models.IntegerField()
+    FirstNumber = models.PositiveIntegerField()
+    SecondNumber = models.PositiveSmallIntegerField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.FirstNumber} & {self.SecondNumber} at {self.Time}'
 
 class primem(models.Model):
-    Number = models.IntegerField()
+    Number = models.PositiveBigIntegerField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.Number} & {self.Time}'
     
 class otpm(models.Model):
-    Number = models.IntegerField()
+    Number = models.PositiveSmallIntegerField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.Number} & {self.Time}'
     
 class fibm(models.Model):
-    Number = models.IntegerField()
+    Number = models.PositiveBigIntegerField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.Number} & {self.Time}'
     
 class sqrtm(models.Model):
-    Number = models.IntegerField()
+    Number = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.Number} & {self.Time}'
     
 class cbrtm(models.Model):
-    Number = models.IntegerField()
+    Number = models.FloatField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.Number} & {self.Time}'
 
 class facm(models.Model):
-    Number = models.IntegerField()
+    Number = models.PositiveSmallIntegerField()
     Time = models.DateTimeField(default = datetime.now())
     
     def __str__(self):
         return f'{self.Number} & {self.Time}'
+
